// src/types/admin.ts
export type ReviewStatus = 'pending' | 'approved' | 'rejected';

export interface Review {
  _id?: string;
  analystId: number;
  analystName: string;
  reviewerName: string;
  userId?: string | null;
  reviewerId?: string | null;
  userProfilePicture?: string | null;
  rating: number;
  comment: string;
  reviewDate: string;
  status: ReviewStatus;
  createdAt: Date | string;
  updatedAt: Date | string;
  approvedAt?: Date | string;
  rejectedAt?: Date | string;
}